PG-1057 add comments
